Fixed Output and Assignment Methods; Added haveVariable method; Unused logicop method
I fixed up my output and assignment statements since I realized they were erroneous when I ran some test cases. They are better than they are before; they seem to work fine, but I will have to do more testing to be sure.

I added a method called haveVariable to see if a variable exists in the symbol table. I have added this method in the output, input, and term sections within the stmt method; but this is by no means a final method. Feel free to look over it; I'm wondering if there's a more efficient way to execute this idea of declaration checking than the one drafted currently.

I also added a comment to the logicop method. According to the grammar, the expr method allows a SIMPLEEXPR to be followed with exactly one LOGICOP and SIMPLEEXPR. However, LOGICOP is not used in the program. I just wanted to point this out; it may be something we have to implement.