Error-Checking
I renamed the method checkSymbol to hasVariable.

I added an if statement to check if there's a line in the middle of the file. It skips over that empty line if it does.

I added an if statement in vdec. Before, even though the var list could be bad, the console still prints that there's no begin statement even though there is one. This should make the error reported, whether in the var list or begin statement, more accurate.

Added comments for error-checking in the simpleexpr and while methods.